+++ tests/test_guests.py
@@ -5,6 +5,7 @@
 # vim: set noai syntax=python ts=4 sw=4:
 """Testing Guests Module and Blueprint Views."""
 
+import pytest
 from flask.testing import FlaskClient
 from werkzeug.test import TestResponse
 
@@ -15,6 +16,28 @@ def test_index(client: FlaskClient) -> None:
     assert response.status_code == 200
     assert b"Guests" in response.data
     assert b"Best Of Only Not My Job Guests" in response.data
+
+
+def test_appearances_by_year(client: FlaskClient) -> None:
+    """Testing guests.routes.appearances_by_year."""
+    response: TestResponse = client.get("/guests/appearances-by-year")
+    assert response.status_code == 200
+    assert b"Appearances by Year" in response.data
+    assert b"Select a Year" in response.data
+
+
+@pytest.mark.parametrize("year", [1998, 2018])
+def test_appearances_by_year_post(client: FlaskClient, year: int) -> None:
+    """Testing guests.routes.appearances_by_year (POST)."""
+    response: TestResponse = client.post(
+        "/guests/appearances-by-year",
+        data={"year": year},
+    )
+    assert response.status_code == 200
+    assert b"Appearances by Year" in response.data
+    assert b"Select a Year" in response.data
+    assert b"Show Date" in response.data
+    assert b"Scoring Exception" in response.data
 
 
 def test_best_of_only_not_my_job_guests(client: FlaskClient) -> None: